things i don't quite like
Removed the redudant cap-add/seccomp params, but i hate running
a priv mode container. It makes my skin crawl. So because of this
I'm watching the work outlined [HERE](https://blog.jessfraz.com/post/building-container-images-securely-on-kubernetes/).

So when possible, I'll remove the ugly priv/net modes. For now,
I'm not quite as smart as some others working to fix this issue.

- v1k